[constraints][3/X] add generics constraints
Adding a set of constraints to all subtyping judgements, and fallout.
No new rules are added, and one admit had to be added into adequacy.

The next diff will add the new subtyping rules to use these constraints
and the necessary checks (New / Class and Method defs), but won't fix
adequacy. The diff after that will tie the knot and fix everything.