[2/2] Updating runtime checks
Adding a new subtype rule to inject `ExT` into `ClassT` when
there is no generic. With this rule, one can do a runtime tag
check with a class that has no generic, and call its methods
in the conditional branch.

We plan to support classes with generics (and probably remove
this rule which would be too specific) at a later stage