[dynamic] Adding two dynamic related rules: if/get
Adding dynamic rules for if (if the condition is : dynamic) and
get (if the receiver is dynamic)

Interesting note: doing the proof showed that we must restrict the
receiver of a dynamic get to avoid performing it on `ThisE`.
Otherwise, we can't be sure the resulting value is correct w.r.t
dynamic.

First (small) victory for Shack ;)