[shack] alternative presentation of `new`
TL;DR: no effect on soundness whatsoever, only on expressivity

In the main branch, the `NewC` command expects full type arguments
to be passed to it. This branch investigate an alternative design
where the `NewC` command doesn't care/need these type parements,
and the developper has to provide them when using the `NewTy` rule.

This should allow more flexibility when re-typing code under the dynamic
conditions, where the type parements for a `new` might not be the same
as the non-dynamic one.