Introducing "exact classes"
A class type can now be exact or inexact. An exact type is exactly the
mentionned classed, not one of its subclass.

For example `$c = new C()` creates a `$c` of exactly the type C. But
in `function f(C $c)`, we could get a subtype of C, not exactly C.

This is a prerequisite to correctly model the `this` type which is an
exact type